run in parellel

+++ Edt.Bond.Migration.Reconciliation.Framework/Output/LoadFileWriter.cs
@@ -9,30 +9,37 @@ namespace Edt.Bond.Migration.Reconciliation.Framework.Output
 
         private string _encap = "\"";
         private string _delimiter = "\t";
+        private bool _shouldOutput;
 
         public StreamWriter _streamWriter;
         public StreamReader _streamReader;
 
         public LoadFileWriter()
         {
-            if(File.Exists(Path.Combine(Settings.ReportingDirectory, $"{_fileStem}.txt")))
-                _streamReader = new StreamReader(Path.Combine(Settings.ReportingDirectory, $"{_fileStem}.txt"));
+            if (Settings.GenerateLoadFile)
+            {
+                if (File.Exists(Path.Combine(Settings.ReportingDirectory, $"{_fileStem}.txt")))
+                    _streamReader = new StreamReader(Path.Combine(Settings.ReportingDirectory, $"{_fileStem}.txt"));
 
-            _streamWriter = new StreamWriter(Path.Combine(Settings.ReportingDirectory, $"{_fileStem}_temp.txt"));
+                _streamWriter = new StreamWriter(Path.Combine(Settings.ReportingDirectory, $"{_fileStem}_temp.txt"));
+
+                _shouldOutput = true;
+            }
         }
         
         public void OutputHeader(string column)
         {
-            var line = _streamReader != null ? $"{_streamReader.ReadLine()}{_delimiter}" : $"DocNumber{_delimiter}";
+            if (_shouldOutput)
+            {
+                var line = _streamReader != null ? $"{_streamReader.ReadLine()}{_delimiter}" : $"DocNumber{_delimiter}";
 
-             _streamWriter.WriteLine($"{line}{column}");
-
+                _streamWriter.WriteLine($"{line}{column}");
+            }
         }
 
         public void OutputRecord(string docNumber, string record)
         {
-
-	        if (Settings.GenerateLoadFile)
+	        if (_shouldOutput)
 	        {
 
 		        var line = _streamReader != null
@@ -55,10 +62,14 @@ namespace Edt.Bond.Migration.Reconciliation.Framework.Output
             _streamReader?.Close();
             _streamWriter?.Close();
 
-            if(File.Exists(Path.Combine(Settings.ReportingDirectory, $"{_fileStem}.txt")))
-                File.Delete(Path.Combine(Settings.ReportingDirectory, $"{_fileStem}.txt"));
+            if (_shouldOutput)
+            {
+                if (File.Exists(Path.Combine(Settings.ReportingDirectory, $"{_fileStem}.txt")))
+                    File.Delete(Path.Combine(Settings.ReportingDirectory, $"{_fileStem}.txt"));
 
-            File.Move(Path.Combine(Settings.ReportingDirectory, $"{_fileStem}_temp.txt"), Path.Combine(Settings.ReportingDirectory, $"{_fileStem}.txt"));
+                File.Move(Path.Combine(Settings.ReportingDirectory, $"{_fileStem}_temp.txt"), Path.Combine(Settings.ReportingDirectory, $"{_fileStem}.txt"));
+
+            }
         }
     }
 }
